third_party/blink/web_tests/svg/custom/animation-values-parsing-error.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests parsing of values attribute.");

var animate = document.createElementNS("http://www.w3.org/2000/svg", "animate");
animate.setAttribute("values", ";;");
animate.setAttribute("values", "300;255;180;30;;");
animate.setAttribute("values", "50%;foo;10%");
animate.setAttribute("values", "50%;;;;10%");
animate.setAttribute("values", "50%;  ;10%");
animate.setAttribute("attributeName", "in");
animate.setAttribute("values", "");

var successfullyParsed = true;
</script>
</body>
</html>
