third_party/blink/web_tests/svg/custom/animation-values-parsing-error-expected.txt

CONSOLE ERROR: line 11: Error: <animate> attribute values: Invalid value, ";;".
CONSOLE ERROR: line 12: Error: <animate> attribute values: Invalid value, "300;255;180;30;;".
CONSOLE ERROR: line 14: Error: <animate> attribute values: Invalid value, "50%;;;;10%".
CONSOLE ERROR: line 15: Error: <animate> attribute values: Invalid value, "50%;  ;10%".
Tests parsing of values attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE
